% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintfr.R
\docType{data}
\name{string_base}
\alias{string_base}
\title{Get a list of base string components for building string formats}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 9 columns.}
\usage{
string_base
}
\description{
Will return information about the base string components that can be used in string_formats.
}
\examples{
string_base
string_format(integer, " ", double)
}
\keyword{datasets}

