% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintfr.R
\docType{data}
\name{string_flag}
\alias{string_flag}
\title{Get a list of optional string format flags}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 7 columns.}
\usage{
string_flag
}
\description{
Each flag is associated with a function. This function will append the flag to an input string format.
    Flag functions take two values: a format to modify, and the value to modify the flag to
    (by default, TRUE). Flags can be looked up by code in the sprintf documentation.
}
\examples{
    string_format(zero_pad(double))
    string_format(zero_pad(double, FALSE))
}
\keyword{datasets}

