% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintfr.R
\name{string_format}
\alias{string_format}
\title{Build a string format.}
\usage{
string_format(..., sep = "")
}
\arguments{
\item{...}{String format pieces}

\item{sep}{String format separator}
}
\description{
Components will be concatenated separated by sep (lists and vectors will be unlisted).
    Raw % signs will be doubled.
    Special elements can be included: string bases, flags, and options.
    See the documentation in string_base, string_flag, and string_option.
}
\examples{
string_format(integer, " ", double)
string_format(integer, double, sep = " ")
}

