% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintfr.R
\name{string_format_}
\alias{string_format_}
\title{Standard evaluation version of string_format}
\usage{
string_format_(args, sep = "")
}
\arguments{
\item{args}{A list of unevaluated time format pieces}

\item{sep}{A separator for pasting pieces together.}
}
\description{
Useful for interactively building time formats.
}
\examples{
base_list = c("double", "integer")
string_format_(list(base_list[1], "' '", base_list[2]) )
}

