% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintfr.R
\docType{data}
\name{string_option}
\alias{string_option}
\title{Get a list of string options}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 3 columns.}
\usage{
string_option
}
\description{
Each string option is associated with a function. This function takes two arguments:
    a string format to be modified and value to modify it with. More information is about each option
    is included in the returned table.
}
\examples{
    string_format(after_decimal(double, 3))
}
\keyword{datasets}

