% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintfr.R
\name{string_table}
\alias{string_table}
\title{Build string format data into handy dataframe form.}
\usage{
string_table(x)
}
\arguments{
\item{x}{A list, dataframe, string, or object that can be converted to a string.}
}
\description{
This is an internal function used for formatting. However, its available
to users because tidy data is best data. It will recursively bind dataframes together rowwise,
converting any non-list to a string, doubling all percentage signs,
and then to dplyr::data_frame(base = "raw", raw_value = .)
}
\examples{
string_table(list(data.frame(base = "raw", raw_value = "1\%\%"),
                 "1\%") )
}

